third_party/blink/web_tests/http/tests/images/document-policy-unoptimized-images-cached-image-expected.txt

CONSOLE ERROR: Document policy violation: Image bpp (byte per pixel) exceeds max value set in lossy-images-max-bpp.
CONSOLE ERROR: Document policy violation: Image bpp (byte per pixel) exceeds max value set in lossy-images-max-bpp.
CONSOLE ERROR: Document policy violation: Image bpp (byte per pixel) exceeds max value set in lossy-images-max-bpp.
